housekeeping

+++ java/GreatProject/src/fi/my/pkg/dependents/Book.java
@@ -31,7 +31,7 @@ public class Book implements Item {
 	
 	protected void validate() {
 		if (id == null) throw new IllegalArgumentException("invalid id");
-		if (id.asLong() < 1) throw new IllegalArgumentException("invalid id " + id);
+		if (id.asInt() < 1) throw new IllegalArgumentException("invalid id " + id);
 		if (!isValidISBN()) throw new IllegalArgumentException("invalid isbn " + isbn);
 	}
